refactor: add development view container and support AB testing
