+++ packages/vscode-extension/src/commandsTreeViewProvider.ts
@@ -4,13 +4,14 @@
 import * as vscode from "vscode";
 import * as path from "path";
 import { ext } from "./extensionVariables";
-import { TreeItem, TreeCategory, Result, FxError, ok, Stage, Func } from "@microsoft/teamsfx-api";
+import { TreeItem, TreeCategory, Result, FxError, ok } from "@microsoft/teamsfx-api";
 import * as StringResources from "./resources/Strings.json";
 import { Correlator } from "@microsoft/teamsfx-core";
 import { Void } from "@microsoft/teamsfx-api";
-import { Commands } from "./controls/Commands";
-import { runCommand, runUserTask } from "./handlers";
-import { TelemetryEvent } from "./telemetry/extTelemetryEvents";
+import { exp } from "./exp";
+import { TreatmentVariables } from "./exp/treatmentVariables";
+import { CommandsWebviewProvider } from "./commandsWebviewProvider";
+import { TreeContainerType } from "./treeview/treeContainerType";
 
 class TreeViewManager {
   private static instance: TreeViewManager;
@@ -85,10 +86,30 @@ class TreeViewManager {
         { name: "edit", custom: false }
       ),
     ];
-    const developmentProvider = new CommandsTreeViewProvider(developmentCommand);
-    disposables.push(
-      vscode.window.registerTreeDataProvider("teamsfx-development", developmentProvider)
-    );
+
+    let developmentProvider: any;
+    if (
+      await exp
+        .getExpService()
+        .getTreatmentVariableAsync(
+          TreatmentVariables.VSCodeConfig,
+          TreatmentVariables.CustomizeTreeview,
+          true
+        )
+    ) {
+      developmentProvider = new CommandsWebviewProvider(TreeContainerType.Development);
+      disposables.push(
+        vscode.window.registerWebviewViewProvider(
+          "teamsfx-development-webview",
+          developmentProvider
+        )
+      );
+    } else {
+      developmentProvider = new CommandsTreeViewProvider(developmentCommand);
+      disposables.push(
+        vscode.window.registerTreeDataProvider("teamsfx-development", developmentProvider)
+      );
+    }
 
     const deployCommand = [
       new TreeViewCommand(
@@ -155,10 +176,28 @@ class TreeViewManager {
         { name: "sync", custom: false }
       ),
     ];
-    const deployProvider = new CommandsTreeViewProvider(deployCommand);
-    disposables.push(vscode.window.registerTreeDataProvider("teamsfx-deployment", deployProvider));
-    // const deployProvider = new CommandsWebviewProvider();
-    // disposables.push(vscode.window.registerWebviewViewProvider("teamsfx-deployment", deployProvider));
+
+    let deployProvider: any;
+    if (
+      await exp
+        .getExpService()
+        .getTreatmentVariableAsync(
+          TreatmentVariables.VSCodeConfig,
+          TreatmentVariables.CustomizeTreeview,
+          true
+        )
+    ) {
+      deployProvider = new CommandsWebviewProvider(TreeContainerType.Deployment);
+      disposables.push(
+        vscode.window.registerWebviewViewProvider("teamsfx-deployment-webview", deployProvider)
+      );
+    } else {
+      deployProvider = new CommandsTreeViewProvider(deployCommand);
+      disposables.push(
+        vscode.window.registerTreeDataProvider("teamsfx-deployment", deployProvider)
+      );
+    }
+
     const helpCommand = [
       new TreeViewCommand(
         StringResources.vsc.commandsTreeViewProvider.quickStartTitle,
@@ -313,92 +352,6 @@ class TreeViewManager {
       value.dispose();
     });
   }
-}
-
-class CommandsWebviewProvider implements vscode.WebviewViewProvider {
-  private _view?: vscode.WebviewView;
-
-  constructor() {}
-
-  public resolveWebviewView(
-    webviewView: vscode.WebviewView,
-    context: vscode.WebviewViewResolveContext,
-    _token: vscode.CancellationToken
-  ) {
-    this._view = webviewView;
-
-    webviewView.webview.options = {
-      // Allow scripts in the webview
-      enableScripts: true,
-    };
-
-    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
-
-    webviewView.webview.onDidReceiveMessage(async (msg) => {
-      switch (msg.command) {
-        case Commands.ExecuteCommand:
-          await vscode.commands.executeCommand(msg.id, "TreeView");
-          break;
-        case Commands.OpenExternalLink:
-          vscode.env.openExternal(vscode.Uri.parse(msg.data));
-      }
-    });
-  }
-
-  private _getHtmlForWebview(webview: vscode.Webview) {
-    const scriptBasePathOnDisk = vscode.Uri.file(path.join(ext.context.extensionPath, "out/"));
-    const scriptBaseUri = scriptBasePathOnDisk.with({ scheme: "vscode-resource" });
-    // Get the local path to main script run in the webview, then convert it to a uri we can use in the webview.
-    const scriptPathOnDisk = vscode.Uri.file(
-      path.join(ext.context.extensionPath, "out/src", "tree.js")
-    );
-    const scriptUri = scriptPathOnDisk.with({ scheme: "vscode-resource" });
-    // // Do the same for the stylesheet.
-    // const styleResetUri = webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, 'media', 'reset.css'));
-    // const styleVSCodeUri = webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, 'media', 'vscode.css'));
-    // const styleMainUri = webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, 'media', 'main.css'));
-    const codiconsUri = webview.asWebviewUri(
-      vscode.Uri.file(
-        path.join(
-          ext.context.extensionPath,
-          "node_modules",
-          "@vscode/codicons",
-          "dist",
-          "codicon.css"
-        )
-      )
-    );
-    // Use a nonce to only allow a specific script to be run.
-    const nonce = getNonce();
-
-    return `<!DOCTYPE html>
-        <html lang="en">
-          <head>
-            <meta charset="UTF-8">
-            <base href='${scriptBaseUri}' />
-            <meta http-equiv="Content-Security-Policy" content="font-src ${webview.cspSource};">
-            <meta name="viewport" content="width=device-width, initial-scale=1.0">
-            <title>ms-teams</title>
-            <link href="${codiconsUri}" rel="stylesheet" />
-          </head>
-          <body style="padding: 0 0">
-            <div id="root"></div>
-            <script>
-              const vscode = acquireVsCodeApi();
-            </script>
-            <script nonce="${nonce}"  type="module" src="${scriptUri}"></script>
-          </body>
-        </html>`;
-  }
-}
-
-function getNonce() {
-  let text = "";
-  const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  for (let i = 0; i < 32; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
 }
 
 export default TreeViewManager.getInstance();
