+++ packages/vscode-extension/src/controls/declarations.d.ts
@@ -8,3 +8,4 @@ declare const vscode: vscode;
 declare const panelType: string;
 declare const isSupportedNode: boolean;
 declare const isMacPlatform: boolean;
+declare const containerType: string;
