+++ packages/vscode-extension/src/exp/treatmentVariables.ts
@@ -5,6 +5,7 @@ export class TreatmentVariables {
   public static readonly ExpandCreateCard = "expandcreatecard";
   public static readonly SidebarWelcome = "sidebarwelcome";
   public static readonly EmbeddedSurvey = "embeddedsurvey";
+  public static readonly CustomizeTreeview = "customizetreeview";
 }
 
 export class TreatmentVariableValue {
