+++ packages/vscode-extension/src/treeview/theme.ts
