+++ packages/vscode-extension/src/treeview/tree.scss
