+++ packages/vscode-extension/src/treeview/treeContainerType.ts
@@ -1,0 +1,4 @@
+export enum TreeContainerType {
+  Development = "development",
+  Deployment = "deployment",
+}
